version beta 1.0

+++ app/Http/Controllers/Profile_EditController.php
@@ -26,9 +26,12 @@ class Profile_EditController extends Controller
     $prev = Auth::user()->img_name;        
 
 
+
     $image_name = '';
     if ($datos->hasFile('img_name')) {
-        unlink("storage/uploads/user$id/$prev");
+        if(Auth::user()->img_name){
+            unlink("storage/uploads/user$id/$prev");
+        }
         $image = $datos->file('img_name');
         $image_name = time().'.'.$image->getClientOriginalExtension();
         $destinationPath = 'storage/uploads/user'.$id;
